Implementation of spatially varying viscosity
This branch implements a spatially varying viscosity. This is useful in
cases where a sponge region is needed. A sponge region is useful to
damp out the gradients in turbulent structures before exiting
a characterstic boundary. The flow field with smaller gradients in space
is less likely to cause reflections when exiting a characterstic
boundary.