This implements an averaging plugin.
This implements an averaging plugin which performs windowed averaging
if the solution every 'nsteps' iterations. The output is written to a
separate file with _avg.pyfrs ending. The writing is skipped every
'writeskip' iterations with respect to the writing of the solutions.
This makes use of the existing infrastructure of file writing, hence the
serial-h5 options under soln-output influences this plugin writer too.