Reader improvements.
This modifies the native reader such that only the master reads the
keys and broadcasts it. All further __iter__ is over this tuple of
of keys,

For this to work, the mpi (when present) needs to be initialized
before the reader is invoked. In case the MPI is not initialized,
every rank reads the keys from the file and caches it.